requirements.txt
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
yfinance>=0.2.35
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0 